+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+### Changed
+- Input connectors now throttle parsing error messages if their share is more than 10% of the parsing attempts.
+
 ## [0.21.0] - 2025-03-19
 
 ### Added